Workaround for RethinkDB 2.0's API change related to `changes` (#23)
RethinkDB 2.0 doesn’t add the `changes` key to a query result if the
document hasn’t been modified. This introduced an annoying bug in
remodel and this commit aims to provide a temporary fix. Follow on the
discussion on #23 for a conclusion on RethinkDB’s API and change this
quick fix to a better one.